backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Redis & cache
redis[hiredis]==5.1.1

# Background tasks
celery==5.4.0
flower==2.0.1

# Auth & security
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12

# Settings & env
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP client
httpx==0.27.2

# Utilities
python-slugify==8.0.4
orjson==3.10.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
